ml_api/requirements.txt
scikit-learn==1.4.1.post1
scipy==1.12.0
xgboost==2.0.3
pandas==2.2.1
Flask==2.0.3
Werkzeug==2.2.2
PyYAML==6.0.1
requests==2.27.1
python-dotenv==1.0.0
mod-wsgi==4.9.4
mod-wsgi-httpd==2.4.54.1
mod-wsgi-standalone==4.9.4